test: move rack_type test into resource_netbox_rack_test.go
Addresses review feedback: TestAccNetboxRack_rackType tests a feature of
netbox_rack, so it belongs in the rack resource's test file rather than a
separate one.
